+++ examples/3D_Light4/source/main.c
@@ -22,7 +22,7 @@ int main(int argc, char **argv) {
 
     GRRLIB_SetBackgroundColour(0x40, 0x40, 0x40, 0xFF);
 
-    while(1) {
+    while(SYS_MainLoop()) {
         GRRLIB_2dMode();
         WPAD_ScanPads();
         if (WPAD_ButtonsDown(0) & WPAD_BUTTON_HOME)  break;