+++ examples/basic_drawing/source/main.c
@@ -58,8 +58,8 @@ int main() {
     u8 FPS = 0;
 
     ir_t ir1;
-    guVector triangle[] = {{400,200,0.0f}, {500,400,0.0f}, {300,400,0.0f}};
-    u32 trianglecolor[] = {GRRLIB_GREEN, GRRLIB_RED, GRRLIB_BLUE};
+    const guVector triangle[] = {{400,200,0.0f}, {500,400,0.0f}, {300,400,0.0f}};
+    const u32 trianglecolor[] = {GRRLIB_GREEN, GRRLIB_RED, GRRLIB_BLUE};
 
     GRRLIB_Init();
 
@@ -90,7 +90,7 @@ int main() {
     GRRLIB_texImg *tex_BMfont5 = GRRLIB_LoadTexture(BMfont5_png);
     GRRLIB_InitTileSet(tex_BMfont5, 8, 16, 0);
 
-    while(1) {
+    while(SYS_MainLoop()) {
         WPAD_SetVRes(0, 640, 480);
         WPAD_ScanPads();
         const u32 wpaddown = WPAD_ButtonsDown(0);