+++ examples/blending/source/main.c
@@ -60,7 +60,7 @@ int main() {
     GRRLIB_SetMidHandle( GFX_Blob[2], true );
 
 
-    while (true) {
+    while (SYS_MainLoop()) {
         WPAD_ScanPads();
         const u32 WPADKeyDown = WPAD_ButtonsDown(WPAD_CHAN_0);
         WPAD_SetVRes(WPAD_CHAN_0, WinW, WinH);
@@ -157,7 +157,9 @@ int main() {
         }
     }
     ExitGame();
-    return 0;
+
+    // Exit application
+    exit(0);
 }
 
 static void ExitGame(void) {
@@ -170,7 +172,4 @@ static void ExitGame(void) {
 
     // Deinitialize GRRLIB & Video
     GRRLIB_Exit();
-
-    // Exit application
-    exit(0);
 }